+++ apps/vscode/src/providers/preview/preview.ts
@@ -112,6 +112,7 @@ export function activatePreview(
   const onSave = async (docUri: Uri) => {
     const editor = findQuartoEditor(
       engine,
+      quartoContext,
       (editorDoc) => editorDoc.uri.fsPath === docUri.fsPath
     );
     if (editor) {
@@ -120,7 +121,7 @@ export function activatePreview(
         (await renderOnSave(engine, editor.document)) &&
         (await previewManager.isPreviewRunningForDoc(editor.document))
       ) {
-        await previewDoc(editor, undefined, true, engine);
+        await previewDoc(editor, undefined, true, engine, quartoContext);
       }
     }
   };
@@ -162,6 +163,7 @@ export async function previewDoc(
   format: string | null | undefined,
   renderOnSave: boolean,
   engine: MarkdownEngine,
+  quartoContext: QuartoContext,
   onShow?: () => void
 ) { 
   // set the slide index from the source editor so we can
@@ -192,6 +194,7 @@ export async function previewDoc(
   // execute the preview (rerefresh the reference after save)
   const previewEditor = findQuartoEditor(
     engine,
+    quartoContext,
     (editorDoc) => editorDoc.uri.fsPath === editor.document.uri.fsPath
   );
   if (previewEditor) {
@@ -610,7 +613,7 @@ class PreviewManager {
 
       // find existing visible instance
       const fileUri = Uri.file(errorLoc.file);
-      const editor = findQuartoEditor(this.engine_, (doc) => doc.uri.fsPath === fileUri.fsPath);
+      const editor = findQuartoEditor(this.engine_, this.quartoContext_, (doc) => doc.uri.fsPath === fileUri.fsPath);
       if (editor) {
         if (editor.textEditor) {
           // if the current selection is outside of the error region then